rename the c api's files
