+++ samples/c/text_generation/chat_sample_c.c
@@ -5,7 +5,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "openvino/genai/c/llm_pipeline_c.h"
+#include "openvino/genai/c/llm_pipeline.h"
 
 #define MAX_PROMPT_LENGTH 64
 #define MAX_OUTPUT_LENGTH 1024
