+++ src/c/include/openvino/genai/c/generation_config.h
