+++ src/c/include/openvino/genai/c/perf_metrics.h
